backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.1.1
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
pillow==11.0.0
pip==25.3
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0